Add files via upload
There are N taxis in a travels company numbered from 1 to N. There are P number of pickup points which are all in a straight line. The distance between these pickup points and the related time to travel in minutes (for both to and fro joumey) is also passed as the input to the program. For each trip, the travels company charges Rs.l for the initial K kilometers and Rs.X for every subsequent kilometer.

- All taxis are position at Pickup point 1 initially.

- When a customer books a taxi, he chooses a pickup point and a free taxi which is closest to that pickup point is assigned for the trip. If more than one taxis are free and are away at equal distance, the taxi with lower revenue is chosen. If the revenue is also same, then the taxi with lower number is chosen.

- The charge is only from the pickup point to the dropping point. (The distance travelled by the taxi till the pickup point is not charged)

- The booking request is accepted only if the distance of a available taxi to the pickup point is less than or equal to Y kilometers.

- If no taxi is free at the time of booking request, then the booking is cancelled.

- If a taxi is free at the time of booking request and the booking is accepted the customer must wait for the additional time taken for the taxi to travel from the current location to the desired pickup point.

The program must print the Taxi number assigned to each customer, amount paid as fare and the drop time
Input Format:

The first line contains N and P separated by a space.

The second line contains P-1 values denoting the distance between the adjacent pickup points. The third line contains P-1 values denoting the time taken in minutes to travel between the adjacent

pickup points. The fourth line contains the values of K, 1, X and Y separated by a space,

The fifth line contains the number of booking requests B.

Next B lines contain the name of the customer, desired pickup point, drop point and the booking request

time in 24 hours format HH:MM